Blog/pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Tariq Zeyad Sharaia <tariq.z.sharaia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
pytz = "^2021.3"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
